chore(server-split): v0.4.x PR-E — Phase 7 self-evolution + learn + patch routes → routes/evolution.py (13 endpoints)
docs/design/v0.4.x-server-split.md 8-PR sequence 여섯 번째 PR.
PR-D 위에 Phase 7 자기진화 + 학습 + 패치 도메인 분리.

## 이동된 surface (~589 lines 서버에서 제거)

### 13 endpoints
- /admin/proposals/ + approve/reject/generate (4)
- /admin/memory/save-rejection (RejectionRequest 동반 이동)
- /admin/evo-reports/
- /admin/learn/topic, /admin/learn/from-errors, /admin/learn/error-queries
- /admin/patches, /admin/patch/approve, /admin/patch/audit, /admin/patch/reject

### 1 Pydantic 모델 + 1 module helper
- RejectionRequest
- _cr_shadow_proposal_create

## server_llmwiki.py 변경
- 13 endpoint + 1 model + 1 helper 제거
- `from routes.evolution import router as evolution_router` + `app.include_router(evolution_router)` (artifacts_router 다음)

## tests/_server_split_helpers.py
- routes.evolution 추가

## tests source-scan 갱신
- test_chat_wiki_save.py / test_evolution_audit_query.py / test_evolution_bench_gate.py / test_self_evolution_gate.py
- combined_server_source() 사용

## 검증

| invariant | 결과 |
|---|---|
| 1. URL byte-identical | ✓ 125 endpoints identical to origin/main |
| 2-5. middleware/audit/static/startup | ✓ smoke pass |
| 6. pytest 회귀 | ✓ **3294 passed, 0 failed** |
| ruff F-class | ✓ All checks passed |

server_llmwiki.py: 4281 → 3692 lines (-589).
8-PR roadmap 진행률 ≈ **37%** (2202/5894).

## Out of scope
- PR-F: /code/*
- PR-G: 잔여 /admin/* (가장 큰 덩어리)
- PR-H: server 최종 정리

Quality delta: exempt (label: chore). URL byte-identical, core/ 무관.
