+++ tests/test_chat_wiki_save.py
@@ -19,7 +19,6 @@ Run:
 """
 from __future__ import annotations
 
-import inspect
 import os
 import re
 import sys
@@ -71,8 +70,8 @@ class PipelineProposalAlwaysCreatedTests(unittest.TestCase):
 class ResponseShapeTests(unittest.TestCase):
     @classmethod
     def setUpClass(cls):
-        import server_llmwiki as srv
-        cls.src = inspect.getsource(srv)
+        from tests._server_split_helpers import combined_server_source
+        cls.src = combined_server_source()
 
     def test_response_model_has_field(self):
         m = re.search(
